refactor: Decompose Settings page into modular section components
Break down the monolithic 1200+ line SettingsPage.tsx into 8 focused section components (GeneralSettings, AppearanceSettings, AIModelSettings, GitSettings, ReviewSettings, AccountSettings, AdvancedSettings, AboutSettings) and a shared SettingsRow component. The main SettingsPage now acts as a lightweight orchestrator.

This improves maintainability, testability, and scalability while preserving all existing functionality.
