+++ src/components/settings/SettingsPage.tsx
@@ -1,97 +1,65 @@
 'use client';
 
-import { useState, useEffect, useCallback } from 'react';
+import { useState } from 'react';
 import { Button } from '@/components/ui/button';
 import { ScrollArea } from '@/components/ui/scroll-area';
-import { Switch } from '@/components/ui/switch';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
 import {
   ArrowLeft,
-  MessageSquare,
-  Palette,
-  GitBranch,
-  FileCode,
-  Bot,
-  User,
-  Beaker,
-  MessageCircle,
-  RefreshCw,
-  FileText,
-  BookOpen,
   Settings2,
-  ExternalLink,
-  CheckCircle2,
+  Palette,
+  Bot,
+  GitBranch,
   Eye,
-  EyeOff,
-  LogOut,
-  FolderOpen,
+  User,
+  Wrench,
+  Info,
 } from 'lucide-react';
 import { cn } from '@/lib/utils';
-import { useSettingsStore } from '@/stores/settingsStore';
-import { openFolderDialog, setMinimizeToTray, requestNotificationPermission } from '@/lib/tauri';
-import { getWorkspacesBasePath, setWorkspacesBasePath, getGlobalReviewPrompts, setGlobalReviewPrompts, getGlobalPRTemplate, setGlobalPRTemplate, getEnvSettings, setEnvSettings, getAnthropicApiKey, setAnthropicApiKey } from '@/lib/api';
-import { EDITOR_THEMES } from '@/lib/monacoThemes';
-import { REVIEW_PROMPTS, REVIEW_TYPE_META } from '@/hooks/useReviewTrigger';
-import { Textarea } from '@/components/ui/textarea';
-import { useToast } from '@/components/ui/toast';
-import { useTheme } from 'next-themes';
-import { useInstalledApps } from '@/hooks/useInstalledApps';
-import { APP_REGISTRY } from '@/lib/openApps';
+import { GeneralSettings } from './sections/GeneralSettings';
+import { AppearanceSettings } from './sections/AppearanceSettings';
+import { AIModelSettings } from './sections/AIModelSettings';
+import { GitSettings } from './sections/GitSettings';
+import { ReviewSettings } from './sections/ReviewSettings';
+import { AccountSettings } from './sections/AccountSettings';
+import { AdvancedSettings } from './sections/AdvancedSettings';
+import { AboutSettings } from './sections/AboutSettings';
 
 interface SettingsPageProps {
   onBack: () => void;
 }
 
 type SettingsCategory =
-  | 'chat'
+  | 'general'
   | 'appearance'
+  | 'ai-models'
   | 'git'
   | 'review'
-  | 'env'
-  | 'claude-code'
   | 'account'
-  | 'experimental'
-  | 'feedback'
-  | 'updates'
-  | 'advanced';
+  | 'advanced'
+  | 'about';
 
 interface NavItem {
   id: SettingsCategory;
   label: string;
   icon: React.ReactNode;
-  external?: boolean;
 }
 
 const mainNavItems: NavItem[] = [
-  { id: 'chat', label: 'Chat', icon: <MessageSquare className="w-3.5 h-3.5" /> },
+  { id: 'general', label: 'General', icon: <Settings2 className="w-3.5 h-3.5" /> },
   { id: 'appearance', label: 'Appearance', icon: <Palette className="w-3.5 h-3.5" /> },
+  { id: 'ai-models', label: 'AI & Models', icon: <Bot className="w-3.5 h-3.5" /> },
   { id: 'git', label: 'Git', icon: <GitBranch className="w-3.5 h-3.5" /> },
-  { id: 'review', label: 'Review', icon: <Eye className="w-3.5 h-3.5" /> },
-  { id: 'env', label: 'Env', icon: <FileCode className="w-3.5 h-3.5" /> },
-  { id: 'claude-code', label: 'Claude Code', icon: <Bot className="w-3.5 h-3.5" /> },
+  { id: 'review', label: 'Review & PRs', icon: <Eye className="w-3.5 h-3.5" /> },
   { id: 'account', label: 'Account', icon: <User className="w-3.5 h-3.5" /> },
 ];
 
 const moreNavItems: NavItem[] = [
-  { id: 'experimental', label: 'Experimental', icon: <Beaker className="w-3.5 h-3.5" /> },
-  { id: 'feedback', label: 'Feedback', icon: <MessageCircle className="w-3.5 h-3.5" /> },
-  { id: 'updates', label: 'Check for updates', icon: <RefreshCw className="w-3.5 h-3.5" /> },
-  { id: 'advanced', label: 'Advanced', icon: <Settings2 className="w-3.5 h-3.5" /> },
-];
-
-const externalLinks = [
-  { label: 'Changelog', icon: <FileText className="w-3.5 h-3.5" />, url: '#' },
-  { label: 'Docs', icon: <BookOpen className="w-3.5 h-3.5" />, url: '#' },
+  { id: 'advanced', label: 'Advanced', icon: <Wrench className="w-3.5 h-3.5" /> },
+  { id: 'about', label: 'About', icon: <Info className="w-3.5 h-3.5" /> },
 ];
 
 export function SettingsPage({ onBack }: SettingsPageProps) {
-  const [selectedCategory, setSelectedCategory] = useState<SettingsCategory>('chat');
+  const [selectedCategory, setSelectedCategory] = useState<SettingsCategory>('general');
 
   return (
     <div className="flex h-full bg-content-background">
@@ -150,19 +118,6 @@ export function SettingsPage({ onBack }: SettingsPageProps) {
                     {item.label}
                   </Button>
                 ))}
-                {externalLinks.map((link) => (
-                  <Button
-                    key={link.label}
-                    variant="ghost"
-                    size="sm"
-                    className="w-full justify-start gap-2 h-7 text-xs"
-                    onClick={() => window.open(link.url, '_blank')}
-                  >
-                    {link.icon}
-                    {link.label}
-                    <ExternalLink className="w-3 h-3 ml-auto text-muted-foreground" />
-                  </Button>
-                ))}
               </div>
             </div>
           </div>
@@ -176,1072 +131,17 @@ export function SettingsPage({ onBack }: SettingsPageProps) {
 
         <ScrollArea className="flex-1">
           <div className="max-w-2xl mx-auto py-8 px-8">
-            {selectedCategory === 'chat' && <ChatSettings />}
+            {selectedCategory === 'general' && <GeneralSettings />}
             {selectedCategory === 'appearance' && <AppearanceSettings />}
+            {selectedCategory === 'ai-models' && <AIModelSettings />}
             {selectedCategory === 'git' && <GitSettings />}
             {selectedCategory === 'review' && <ReviewSettings />}
-            {selectedCategory === 'env' && <EnvSettings />}
-            {selectedCategory === 'claude-code' && <ClaudeCodeSettings />}
             {selectedCategory === 'account' && <AccountSettings />}
-            {selectedCategory === 'experimental' && <ExperimentalSettings />}
-            {selectedCategory === 'feedback' && <FeedbackSettings />}
-            {selectedCategory === 'updates' && <UpdatesSettings />}
             {selectedCategory === 'advanced' && <AdvancedSettings />}
+            {selectedCategory === 'about' && <AboutSettings />}
           </div>
         </ScrollArea>
       </div>
-    </div>
-  );
-}
-
-// Settings Row Component
-function SettingsRow({
-  title,
-  description,
-  children,
-}: {
-  title: string;
-  description?: string;
-  children: React.ReactNode;
-}) {
-  return (
-    <div className="flex items-center justify-between py-3 border-b border-border/50">
-      <div className="flex-1 pr-4">
-        <h4 className="text-sm font-medium">{title}</h4>
-        {description && (
-          <p className="text-xs text-muted-foreground mt-0.5">{description}</p>
-        )}
-      </div>
-      <div className="shrink-0">{children}</div>
-    </div>
-  );
-}
-
-// Chat Settings
-function ChatSettings() {
-  const confirmCloseActiveTab = useSettingsStore((s) => s.confirmCloseActiveTab);
-  const setConfirmCloseActiveTab = useSettingsStore((s) => s.setConfirmCloseActiveTab);
-  const showTokenUsage = useSettingsStore((s) => s.showTokenUsage);
-  const setShowTokenUsage = useSettingsStore((s) => s.setShowTokenUsage);
-  const desktopNotifications = useSettingsStore((s) => s.desktopNotifications);
-  const setDesktopNotifications = useSettingsStore((s) => s.setDesktopNotifications);
-  const defaultModel = useSettingsStore((s) => s.defaultModel);
-  const setDefaultModel = useSettingsStore((s) => s.setDefaultModel);
-  const defaultThinking = useSettingsStore((s) => s.defaultThinking);
-  const setDefaultThinking = useSettingsStore((s) => s.setDefaultThinking);
-  const reviewModel = useSettingsStore((s) => s.reviewModel);
-  const setReviewModel = useSettingsStore((s) => s.setReviewModel);
-  const defaultPlanMode = useSettingsStore((s) => s.defaultPlanMode);
-  const setDefaultPlanMode = useSettingsStore((s) => s.setDefaultPlanMode);
-  const soundEffects = useSettingsStore((s) => s.soundEffects);
-  const setSoundEffects = useSettingsStore((s) => s.setSoundEffects);
-  const soundEffectType = useSettingsStore((s) => s.soundEffectType);
-  const setSoundEffectType = useSettingsStore((s) => s.setSoundEffectType);
-  const sendWithEnter = useSettingsStore((s) => s.sendWithEnter);
-  const setSendWithEnter = useSettingsStore((s) => s.setSendWithEnter);
-  const autoConvertLongText = useSettingsStore((s) => s.autoConvertLongText);
-  const setAutoConvertLongText = useSettingsStore((s) => s.setAutoConvertLongText);
-  const defaultOpenApp = useSettingsStore((s) => s.defaultOpenApp);
-  const setDefaultOpenApp = useSettingsStore((s) => s.setDefaultOpenApp);
-  const { installedApps } = useInstalledApps();
-
-  const handleNotificationToggle = useCallback(async (enabled: boolean) => {
-    if (enabled) {
-      const perm = await requestNotificationPermission();
-      if (perm !== 'granted') return;
-    }
-    setDesktopNotifications(enabled);
-  }, [setDesktopNotifications]);
-
-  return (
-    <div>
-      <h2 className="text-xl font-semibold mb-5">Chat</h2>
-
-      <SettingsRow title="Default model" description="Model for new chats">
-        <div className="flex gap-2">
-          <Select value={defaultModel} onValueChange={setDefaultModel}>
-            <SelectTrigger className="w-52">
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="claude-opus-4-6">Claude Opus 4.6</SelectItem>
-              <SelectItem value="claude-sonnet-4-5-20250929">Claude Sonnet 4.5</SelectItem>
-              <SelectItem value="claude-haiku-4-5-20251001">Claude Haiku 4.5</SelectItem>
-            </SelectContent>
-          </Select>
-          <Select
-            value={defaultThinking ? 'thinking-on' : 'thinking-off'}
-            onValueChange={(v) => setDefaultThinking(v === 'thinking-on')}
-          >
-            <SelectTrigger className="w-32">
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="thinking-on">Thinking on</SelectItem>
-              <SelectItem value="thinking-off">Thinking off</SelectItem>
-            </SelectContent>
-          </Select>
-        </div>
-      </SettingsRow>
-
-      <SettingsRow title="Review model" description="Model for code reviews">
-        <div className="flex gap-2">
-          <Select value={reviewModel} onValueChange={setReviewModel}>
-            <SelectTrigger className="w-52">
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="claude-opus-4-6">Claude Opus 4.6</SelectItem>
-              <SelectItem value="claude-sonnet-4-5-20250929">Claude Sonnet 4.5</SelectItem>
-              <SelectItem value="claude-haiku-4-5-20251001">Claude Haiku 4.5</SelectItem>
-            </SelectContent>
-          </Select>
-        </div>
-      </SettingsRow>
-
-      <SettingsRow title="Default Open in App" description="App used by the toolbar Open button">
-        <Select value={defaultOpenApp} onValueChange={setDefaultOpenApp}>
-          <SelectTrigger className="w-44">
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            {(installedApps.length > 0 ? installedApps : APP_REGISTRY).map((app) => (
-              <SelectItem key={app.id} value={app.id}>
-                {app.name}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-      </SettingsRow>
-
-      <SettingsRow
-        title="Default to plan mode"
-        description="Start new chats in plan mode (Claude only)"
-      >
-        <Switch checked={defaultPlanMode} onCheckedChange={setDefaultPlanMode} />
-      </SettingsRow>
-
-      <SettingsRow
-        title="Desktop notifications"
-        description="Get notified when AI finishes working in a chat"
-      >
-        <Switch checked={desktopNotifications} onCheckedChange={handleNotificationToggle} />
-      </SettingsRow>
-
-      <SettingsRow
-        title="Sound effects"
-        description="Play a sound when AI finishes working in a chat"
-      >
-        <div className="flex items-center gap-2">
-          <Select value={soundEffectType} onValueChange={setSoundEffectType}>
-            <SelectTrigger className="w-24">
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="chime">Chime</SelectItem>
-              <SelectItem value="ding">Ding</SelectItem>
-              <SelectItem value="pop">Pop</SelectItem>
-            </SelectContent>
-          </Select>
-          <Button variant="outline" size="sm">
-            Test
-          </Button>
-          <Switch checked={soundEffects} onCheckedChange={setSoundEffects} />
-        </div>
-      </SettingsRow>
-
-      <SettingsRow
-        title="Send messages with"
-        description="Choose which key combination sends messages. Use ⇧↵ for new lines"
-      >
-        <Select
-          value={sendWithEnter ? 'enter' : 'cmd-enter'}
-          onValueChange={(v) => setSendWithEnter(v === 'enter')}
-        >
-          <SelectTrigger className="w-28">
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            <SelectItem value="enter">Enter</SelectItem>
-            <SelectItem value="cmd-enter">⌘ Enter</SelectItem>
-          </SelectContent>
-        </Select>
-      </SettingsRow>
-
-      <SettingsRow
-        title="Auto-convert long text"
-        description="Convert pasted text over 5000 characters into text attachments"
-      >
-        <Switch checked={autoConvertLongText} onCheckedChange={setAutoConvertLongText} />
-      </SettingsRow>
-
-      <SettingsRow
-        title="Show token usage"
-        description="Display token counts and cost breakdown in run summaries"
-      >
-        <Switch
-          checked={showTokenUsage}
-          onCheckedChange={setShowTokenUsage}
-        />
-      </SettingsRow>
-
-      <SettingsRow
-        title="Confirm before closing active conversations"
-        description="Ask for confirmation when closing a conversation with messages"
-      >
-        <Switch
-          checked={confirmCloseActiveTab}
-          onCheckedChange={setConfirmCloseActiveTab}
-        />
-      </SettingsRow>
-    </div>
-  );
-}
-
-// Appearance settings
-function AppearanceSettings() {
-  const editorTheme = useSettingsStore((s) => s.editorTheme);
-  const setEditorTheme = useSettingsStore((s) => s.setEditorTheme);
-  const fontSize = useSettingsStore((s) => s.fontSize);
-  const setFontSize = useSettingsStore((s) => s.setFontSize);
-  const { theme, setTheme } = useTheme();
-
-  // Group themes by light/dark
-  const darkThemes = EDITOR_THEMES.filter((t) => t.isDark);
-  const lightThemes = EDITOR_THEMES.filter((t) => !t.isDark);
-
-  return (
-    <div>
-      <h2 className="text-xl font-semibold mb-5">Appearance</h2>
-      <SettingsRow title="Theme" description="Choose your preferred theme">
-        <Select value={theme} onValueChange={setTheme}>
-          <SelectTrigger className="w-32">
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            <SelectItem value="system">System</SelectItem>
-            <SelectItem value="light">Light</SelectItem>
-            <SelectItem value="dark">Dark</SelectItem>
-          </SelectContent>
-        </Select>
-      </SettingsRow>
-      <SettingsRow title="Editor theme" description="Syntax highlighting theme for code editors">
-        <Select value={editorTheme} onValueChange={setEditorTheme}>
-          <SelectTrigger className="w-44">
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            <div className="px-2 py-1.5 text-xs font-medium text-muted-foreground">Dark</div>
-            {darkThemes.map((theme) => (
-              <SelectItem key={theme.id} value={theme.id}>
-                {theme.name}
-              </SelectItem>
-            ))}
-            <div className="px-2 py-1.5 text-xs font-medium text-muted-foreground border-t mt-1 pt-2">Light</div>
-            {lightThemes.map((theme) => (
-              <SelectItem key={theme.id} value={theme.id}>
-                {theme.name}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-      </SettingsRow>
-      <SettingsRow title="Font size" description="Adjust the interface font size">
-        <Select value={fontSize} onValueChange={setFontSize}>
-          <SelectTrigger className="w-32">
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            <SelectItem value="small">Small</SelectItem>
-            <SelectItem value="medium">Medium</SelectItem>
-            <SelectItem value="large">Large</SelectItem>
-          </SelectContent>
-        </Select>
-      </SettingsRow>
-    </div>
-  );
-}
-
-function GitSettings() {
-  const branchPrefixType = useSettingsStore((s) => s.branchPrefixType);
-  const setBranchPrefixType = useSettingsStore((s) => s.setBranchPrefixType);
-  const customPrefix = useSettingsStore((s) => s.branchPrefixCustom);
-  const setCustomPrefix = useSettingsStore((s) => s.setBranchPrefixCustom);
-  const deleteBranchOnArchive = useSettingsStore((s) => s.deleteBranchOnArchive);
-  const setDeleteBranchOnArchive = useSettingsStore((s) => s.setDeleteBranchOnArchive);
-  const archiveOnMerge = useSettingsStore((s) => s.archiveOnMerge);
-  const setArchiveOnMerge = useSettingsStore((s) => s.setArchiveOnMerge);
-
-  // TODO: Get actual GitHub username from git config or API
-  const githubUsername = 'mcastilho';
-
-  return (
-    <div>
-      <h2 className="text-xl font-semibold mb-5">Git</h2>
-
-      {/* Branch name prefix */}
-      <div className="py-4 border-b border-border/50">
-        <h4 className="text-sm font-medium">Branch name prefix</h4>
-        <p className="text-sm text-muted-foreground mt-0.5">
-          Prefix for new workspace branch names. Will be followed by a slash.
-        </p>
-
-        <div className="mt-4 space-y-2">
-          <label className="flex items-center gap-3 cursor-pointer">
-            <input
-              type="radio"
-              name="branchPrefix"
-              checked={branchPrefixType === 'github'}
-              onChange={() => setBranchPrefixType('github')}
-              className="w-4 h-4 text-primary border-muted-foreground/50 focus:ring-primary"
-            />
-            <span className="text-sm">GitHub username ({githubUsername})</span>
-          </label>
-
-          <label className="flex items-center gap-3 cursor-pointer">
-            <input
-              type="radio"
-              name="branchPrefix"
-              checked={branchPrefixType === 'custom'}
-              onChange={() => setBranchPrefixType('custom')}
-              className="w-4 h-4 text-primary border-muted-foreground/50 focus:ring-primary"
-            />
-            <span className="text-sm">Custom</span>
-          </label>
-
-          {branchPrefixType === 'custom' && (
-            <div className="ml-7">
-              <input
-                type="text"
-                value={customPrefix}
-                onChange={(e) => setCustomPrefix(e.target.value)}
-                placeholder="Enter custom prefix"
-                className="w-48 px-3 py-1.5 text-sm bg-background border border-border rounded-md focus:outline-none focus:ring-1 focus:ring-primary"
-              />
-            </div>
-          )}
-
-          <label className="flex items-center gap-3 cursor-pointer">
-            <input
-              type="radio"
-              name="branchPrefix"
-              checked={branchPrefixType === 'none'}
-              onChange={() => setBranchPrefixType('none')}
-              className="w-4 h-4 text-primary border-muted-foreground/50 focus:ring-primary"
-            />
-            <span className="text-sm">None</span>
-          </label>
-        </div>
-      </div>
-
-      {/* Delete branch on archive */}
-      <div className="flex items-start justify-between py-4 border-b border-border/50">
-        <div className="flex-1 pr-4">
-          <h4 className="text-sm font-medium">Delete branch on archive</h4>
-          <p className="text-sm text-muted-foreground mt-0.5">
-            Delete the local branch when archiving a workspace.
-            <br />
-            To delete the remote branch,{' '}
-            <a
-              href="https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/configuring-pull-request-merges/managing-the-automatic-deletion-of-branches"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="text-primary hover:underline"
-            >
-              configure it on GitHub
-            </a>
-            .
-          </p>
-        </div>
-        <div className="shrink-0 pt-1">
-          <Switch
-            checked={deleteBranchOnArchive}
-            onCheckedChange={setDeleteBranchOnArchive}
-          />
-        </div>
-      </div>
-
-      {/* Archive on merge */}
-      <div className="flex items-start justify-between py-4 border-b border-border/50">
-        <div className="flex-1 pr-4">
-          <h4 className="text-sm font-medium">Archive on merge</h4>
-          <p className="text-sm text-muted-foreground mt-0.5">
-            Automatically archive a workspace after merging its PR in ChatML
-          </p>
-        </div>
-        <div className="shrink-0 pt-1">
-          <Switch
-            checked={archiveOnMerge}
-            onCheckedChange={setArchiveOnMerge}
-          />
-        </div>
-      </div>
-
-      {/* Confirm archive with uncommitted changes */}
-      <ConfirmArchiveSetting />
-    </div>
-  );
-}
-
-function ConfirmArchiveSetting() {
-  const confirmArchiveDirtySession = useSettingsStore((s) => s.confirmArchiveDirtySession);
-  const setConfirmArchiveDirtySession = useSettingsStore((s) => s.setConfirmArchiveDirtySession);
-
-  return (
-    <div className="flex items-start justify-between py-4 border-b border-border/50">
-      <div className="flex-1 pr-4">
-        <h4 className="text-sm font-medium">Confirm archive with uncommitted changes</h4>
-        <p className="text-sm text-muted-foreground mt-0.5">
-          Show a confirmation dialog when archiving a session that has uncommitted or unpushed changes
-        </p>
-      </div>
-      <div className="shrink-0 pt-1">
-        <Switch
-          checked={confirmArchiveDirtySession}
-          onCheckedChange={setConfirmArchiveDirtySession}
-        />
-      </div>
-    </div>
-  );
-}
-
-function EnvSettings() {
-  const [envVars, setEnvVarsLocal] = useState('');
-  const [savedEnvVars, setSavedEnvVars] = useState('');
-  const [saving, setSaving] = useState(false);
-  const [saveError, setSaveError] = useState<string | null>(null);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    getEnvSettings()
-      .then((vars) => {
-        setEnvVarsLocal(vars);
-        setSavedEnvVars(vars);
-      })
-      .catch((err) => {
-        console.error('Failed to load env settings:', err);
-      })
-      .finally(() => setLoading(false));
-  }, []);
-
-  const hasUnsavedChanges = envVars !== savedEnvVars;
-
-  const handleSave = useCallback(async () => {
-    if (!hasUnsavedChanges) return;
-    setSaving(true);
-    setSaveError(null);
-    try {
-      await setEnvSettings(envVars);
-      setSavedEnvVars(envVars);
-    } catch (err) {
-      setSaveError(err instanceof Error ? err.message : 'Failed to save');
-    } finally {
-      setSaving(false);
-    }
-  }, [envVars, hasUnsavedChanges]);
-
-  return (
-    <div>
-      <h2 className="text-xl font-semibold mb-2">Environment variables</h2>
-      <p className="text-sm text-muted-foreground mb-6">
-        Useful for using third-party providers like Bedrock or Vertex.
-        Changes take effect for new agent sessions.
-      </p>
-
-      {/* Claude Code section */}
-      <div className="flex items-center justify-between mb-3">
-        <h3 className="text-lg font-semibold">Claude Code</h3>
-        <a
-          href="https://docs.anthropic.com/en/docs/claude-code"
-          target="_blank"
-          rel="noopener noreferrer"
-          className="text-sm text-muted-foreground hover:text-foreground flex items-center gap-1"
-        >
-          View docs
-          <ExternalLink className="w-3 h-3" />
-        </a>
-      </div>
-
-      {/* Environment variables textarea */}
-      <textarea
-        value={envVars}
-        onChange={(e) => setEnvVarsLocal(e.target.value)}
-        disabled={loading}
-        className="w-full h-64 px-4 py-3 font-mono text-sm bg-muted/50 border border-border rounded-lg focus:outline-none focus:ring-1 focus:ring-primary resize-none disabled:opacity-50"
-        placeholder="VAR_NAME=value"
-      />
-
-      {/* Footer with format hint and save button */}
-      <div className="flex items-center justify-between mt-4">
-        <div>
-          <p className="text-sm text-muted-foreground">
-            One per line, format:{' '}
-            <code className="px-1.5 py-0.5 bg-muted rounded text-xs font-mono">VAR_NAME=value</code>
-            {' '}or{' '}
-            <code className="px-1.5 py-0.5 bg-muted rounded text-xs font-mono">export VAR_NAME=value</code>
-          </p>
-          {saveError && (
-            <p className="text-sm text-destructive mt-1">{saveError}</p>
-          )}
-        </div>
-        <Button
-          variant="secondary"
-          size="sm"
-          disabled={!hasUnsavedChanges || saving}
-          onClick={handleSave}
-        >
-          {saving ? 'Saving...' : 'Save'}
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-function ClaudeCodeSettings() {
-  const maxThinkingTokens = useSettingsStore((s) => s.maxThinkingTokens);
-  const setMaxThinkingTokens = useSettingsStore((s) => s.setMaxThinkingTokens);
-  const autoApproveSafeCommands = useSettingsStore((s) => s.autoApproveSafeCommands);
-  const setAutoApproveSafeCommands = useSettingsStore((s) => s.setAutoApproveSafeCommands);
-
-  const [apiKeyInput, setApiKeyInput] = useState('');
-  const [apiKeyConfigured, setApiKeyConfigured] = useState(false);
-  const [apiKeyMasked, setApiKeyMasked] = useState('');
-  const [showApiKey, setShowApiKey] = useState(false);
-  const [apiKeySaving, setApiKeySaving] = useState(false);
-  const toasts = useToast();
-
-  useEffect(() => {
-    getAnthropicApiKey().then((data) => {
-      setApiKeyConfigured(data.configured);
-      setApiKeyMasked(data.maskedKey);
-    }).catch(() => {
-      // ignore — settings page should still render
-    });
-  }, []);
-
-  const handleSaveApiKey = async () => {
-    setApiKeySaving(true);
-    try {
-      const result = await setAnthropicApiKey(apiKeyInput);
-      setApiKeyConfigured(result.configured);
-      setApiKeyMasked(result.maskedKey);
-      setApiKeyInput('');
-      toasts.success('New sessions will use this key.', 'API key saved');
-    } catch {
-      toasts.error('Failed to save API key');
-    } finally {
-      setApiKeySaving(false);
-    }
-  };
-
-  const handleRemoveApiKey = async () => {
-    setApiKeySaving(true);
-    try {
-      await setAnthropicApiKey('');
-      setApiKeyConfigured(false);
-      setApiKeyMasked('');
-      toasts.success('API key removed');
-    } catch {
-      toasts.error('Failed to remove API key');
-    } finally {
-      setApiKeySaving(false);
-    }
-  };
-
-  return (
-    <div>
-      <h2 className="text-xl font-semibold mb-5">Claude Code</h2>
-      <SettingsRow
-        title="Auto-approve safe commands"
-        description="Automatically approve read-only commands"
-      >
-        <Switch checked={autoApproveSafeCommands} onCheckedChange={setAutoApproveSafeCommands} />
-      </SettingsRow>
-      <SettingsRow
-        title="Max thinking tokens"
-        description="Maximum tokens for extended thinking"
-      >
-        <Select
-          value={maxThinkingTokens.toString()}
-          onValueChange={(value) => {
-            const n = parseInt(value, 10);
-            if (!isNaN(n) && n > 0) setMaxThinkingTokens(n);
-          }}
-        >
-          <SelectTrigger className="w-32">
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            <SelectItem value="8000">8,000</SelectItem>
-            <SelectItem value="10000">10,000</SelectItem>
-            <SelectItem value="16000">16,000</SelectItem>
-            <SelectItem value="32000">32,000</SelectItem>
-          </SelectContent>
-        </Select>
-      </SettingsRow>
-
-      {/* Anthropic API Key */}
-      <div className="py-4 border-b border-border/50">
-        <h4 className="text-sm font-medium">Anthropic API Key</h4>
-        <p className="text-xs text-muted-foreground mt-0.5">
-          Set an API key to bypass OAuth authentication. Get one from{' '}
-          <a href="https://console.anthropic.com" target="_blank" rel="noopener noreferrer" className="underline hover:text-foreground">
-            console.anthropic.com
-          </a>.
-        </p>
-
-        {apiKeyConfigured && (
-          <p className="text-xs text-muted-foreground mt-2">
-            Current key: <code className="text-xs bg-muted px-1 py-0.5 rounded">{apiKeyMasked}</code>
-          </p>
-        )}
-
-        <div className="flex items-center gap-2 mt-3">
-          <div className="relative flex-1 max-w-xs">
-            <input
-              type={showApiKey ? 'text' : 'password'}
-              value={apiKeyInput}
-              onChange={(e) => setApiKeyInput(e.target.value)}
-              placeholder={apiKeyConfigured ? 'Enter new key to replace' : 'sk-ant-...'}
-              className="w-full px-3 py-1.5 pr-8 text-sm bg-background border border-border rounded-md focus:outline-none focus:ring-1 focus:ring-primary"
-            />
-            <button
-              type="button"
-              onClick={() => setShowApiKey(!showApiKey)}
-              className="absolute right-2 top-1/2 -translate-y-1/2 text-muted-foreground hover:text-foreground"
-            >
-              {showApiKey ? <EyeOff className="w-3.5 h-3.5" /> : <Eye className="w-3.5 h-3.5" />}
-            </button>
-          </div>
-          <Button
-            size="sm"
-            disabled={!apiKeyInput.trim() || apiKeySaving}
-            onClick={handleSaveApiKey}
-          >
-            {apiKeySaving ? 'Saving...' : 'Save'}
-          </Button>
-          {apiKeyConfigured && (
-            <Button
-              size="sm"
-              variant="outline"
-              disabled={apiKeySaving}
-              onClick={handleRemoveApiKey}
-            >
-              Remove
-            </Button>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function AccountSettings() {
-  const [showToken, setShowToken] = useState(false);
-  const [token, setToken] = useState('ghp_xxxxxxxxxxxxxxxxxxxx');
-  const strictPrivacy = useSettingsStore((s) => s.strictPrivacy);
-  const setStrictPrivacy = useSettingsStore((s) => s.setStrictPrivacy);
-
-  // TODO: Get actual user data from auth
-  const user = {
-    name: 'Marcio Castilho',
-    email: 'marcio.castilho@gmail.com',
-    github: 'mcastilho',
-    avatar: 'https://avatars.githubusercontent.com/u/181635?v=4',
-  };
-
-  return (
-    <div>
-      <h2 className="text-xl font-semibold mb-5">Account</h2>
-
-      {/* User profile */}
-      <div className="flex items-center gap-4 pb-6 border-b border-border/50">
-        {/* eslint-disable-next-line @next/next/no-img-element */}
-        <img
-          src={user.avatar}
-          alt={user.name}
-          className="w-16 h-16 rounded-full"
-        />
-        <div>
-          <h3 className="text-lg font-semibold">{user.name}</h3>
-          <p className="text-sm text-muted-foreground">{user.email}</p>
-          <div className="flex items-center gap-1.5 mt-1 text-sm text-muted-foreground">
-            <svg className="w-4 h-4" viewBox="0 0 24 24" fill="currentColor">
-              <path d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z"/>
-            </svg>
-            {user.github}
-          </div>
-        </div>
-      </div>
-
-      {/* Linear Integration */}
-      <div className="flex items-start justify-between py-6 border-b border-border/50">
-        <div>
-          <div className="flex items-center gap-2">
-            <h4 className="text-sm font-medium">Linear Integration</h4>
-            <CheckCircle2 className="w-4 h-4 text-text-success" />
-          </div>
-          <p className="text-sm text-muted-foreground mt-1">Linear is linked.</p>
-        </div>
-        <Button variant="outline" size="sm" className="gap-1.5">
-          Manage
-          <ExternalLink className="w-3 h-3" />
-        </Button>
-      </div>
-
-      {/* GitHub CLI Integration */}
-      <div className="py-6 border-b border-border/50">
-        <div className="flex items-center gap-2">
-          <h4 className="text-sm font-medium">GitHub CLI Integration</h4>
-          <CheckCircle2 className="w-4 h-4 text-text-success" />
-        </div>
-        <p className="text-sm text-muted-foreground mt-1">
-          GitHub CLI is authenticated and ready.
-        </p>
-        <p className="text-sm text-muted-foreground">
-          Signed in as {user.github}
-        </p>
-
-        <div className="mt-4">
-          <label className="text-sm text-muted-foreground">
-            Personal access token (optional)
-          </label>
-          <div className="flex items-center gap-2 mt-1.5">
-            <input
-              type={showToken ? 'text' : 'password'}
-              value={token}
-              onChange={(e) => setToken(e.target.value)}
-              placeholder="ghp_xxxxxxxxxxxxxxxxxxxx"
-              className="flex-1 px-3 py-2 text-sm bg-muted/50 border border-border rounded-md focus:outline-none focus:ring-1 focus:ring-primary"
-            />
-            <Button
-              variant="ghost"
-              size="icon"
-              className="h-9 w-9"
-              onClick={() => setShowToken(!showToken)}
-            >
-              {showToken ? (
-                <EyeOff className="w-4 h-4" />
-              ) : (
-                <Eye className="w-4 h-4" />
-              )}
-            </Button>
-          </div>
-        </div>
-      </div>
-
-      {/* Strict data privacy */}
-      <div className="flex items-start justify-between py-6 border-b border-border/50">
-        <div className="flex-1 pr-4">
-          <h4 className="text-sm font-medium">Strict data privacy</h4>
-          <p className="text-sm text-muted-foreground mt-1">
-            Disable features requiring external AI providers, such as AI generated chat titles.
-          </p>
-        </div>
-        <div className="shrink-0 pt-1">
-          <Switch
-            checked={strictPrivacy}
-            onCheckedChange={setStrictPrivacy}
-          />
-        </div>
-      </div>
-
-      {/* Replay Welcome Tour */}
-      <div className="flex items-start justify-between py-6 border-b border-border/50">
-        <div className="flex-1 pr-4">
-          <h4 className="text-sm font-medium">Welcome tour</h4>
-          <p className="text-sm text-muted-foreground mt-1">
-            Replay the onboarding wizard and guided tour.
-          </p>
-        </div>
-        <Button
-          variant="outline"
-          size="sm"
-          onClick={() => {
-            useSettingsStore.getState().resetOnboarding();
-            window.dispatchEvent(new CustomEvent('close-settings'));
-          }}
-        >
-          Replay Tour
-        </Button>
-      </div>
-
-      {/* Sign out */}
-      <div className="flex items-center justify-between py-6">
-        <h4 className="text-sm font-medium">Sign out</h4>
-        <Button variant="outline" size="sm" className="gap-1.5">
-          <LogOut className="w-4 h-4" />
-          Sign out
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-function ExperimentalSettings() {
-  const parallelAgents = useSettingsStore((s) => s.parallelAgents);
-  const setParallelAgents = useSettingsStore((s) => s.setParallelAgents);
-
-  return (
-    <div>
-      <h2 className="text-xl font-semibold mb-5">Experimental</h2>
-      <p className="text-sm text-muted-foreground mb-4">
-        These features are experimental and may change or be removed.
-      </p>
-      <SettingsRow
-        title="Parallel agents"
-        description="Enable multiple agents working in parallel"
-      >
-        <Switch checked={parallelAgents} onCheckedChange={setParallelAgents} />
-      </SettingsRow>
-    </div>
-  );
-}
-
-function FeedbackSettings() {
-  return (
-    <div>
-      <h2 className="text-xl font-semibold mb-5">Feedback</h2>
-      <p className="text-sm text-muted-foreground">
-        Help us improve ChatML by sharing your feedback.
-      </p>
-    </div>
-  );
-}
-
-function UpdatesSettings() {
-  return (
-    <div>
-      <h2 className="text-xl font-semibold mb-5">Updates</h2>
-      <p className="text-sm text-muted-foreground">
-        ChatML is up to date.
-      </p>
-    </div>
-  );
-}
-
-function ReviewSettings() {
-  const [prompts, setPrompts] = useState<Record<string, string>>({});
-  const [saved, setSaved] = useState<Record<string, string>>({});
-  const [prTemplate, setPRTemplate] = useState('');
-  const [savedPRTemplate, setSavedPRTemplate] = useState('');
-  const [saving, setSaving] = useState(false);
-  const { error: showError } = useToast();
-
-  useEffect(() => {
-    Promise.all([getGlobalReviewPrompts(), getGlobalPRTemplate()])
-      .then(([reviewData, prData]) => {
-        setPrompts(reviewData);
-        setSaved(reviewData);
-        setPRTemplate(prData);
-        setSavedPRTemplate(prData);
-      })
-      .catch(() => {
-        showError('Failed to load settings');
-      });
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const hasChanges = JSON.stringify(prompts) !== JSON.stringify(saved) || prTemplate !== savedPRTemplate;
-
-  const handleSave = async () => {
-    setSaving(true);
-    try {
-      // Filter out empty strings before saving
-      const cleaned: Record<string, string> = {};
-      for (const [k, v] of Object.entries(prompts)) {
-        if (v.trim()) cleaned[k] = v.trim();
-      }
-      await Promise.all([
-        setGlobalReviewPrompts(cleaned),
-        setGlobalPRTemplate(prTemplate.trim()),
-      ]);
-      setPrompts(cleaned);
-      setSaved(cleaned);
-      setPRTemplate(prTemplate.trim());
-      setSavedPRTemplate(prTemplate.trim());
-    } catch {
-      showError('Failed to save settings');
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  return (
-    <div>
-      <h2 className="text-xl font-semibold mb-1">Review</h2>
-      <p className="text-sm text-muted-foreground mb-6">
-        Add custom instructions that will be appended to each review type&apos;s default prompt.
-        These apply globally. Per-workspace overrides can be set in workspace settings.
-      </p>
-
-      <div className="space-y-5">
-        {REVIEW_TYPE_META.map(({ key, label, placeholder }) => (
-          <div key={key}>
-            <label className="text-sm font-medium block mb-1.5">{label}</label>
-            <p className="text-xs text-muted-foreground mb-1.5 line-clamp-1">
-              Default: {REVIEW_PROMPTS[key]?.slice(0, 80)}…
-            </p>
-            <Textarea
-              className="text-sm min-h-[60px]"
-              placeholder={placeholder}
-              value={prompts[key] || ''}
-              onChange={(e) => setPrompts((prev) => ({ ...prev, [key]: e.target.value }))}
-            />
-          </div>
-        ))}
-      </div>
-
-      <div className="mt-8 pt-8 border-t border-border/50">
-        <h3 className="text-lg font-semibold mb-1">PR Creation</h3>
-        <p className="text-sm text-muted-foreground mb-4">
-          Add custom instructions for AI-generated PR descriptions. These apply globally.
-          Per-workspace overrides can be set in workspace settings.
-        </p>
-
-        <label className="text-sm font-medium block mb-1.5">PR Description Prompt</label>
-        <p className="text-xs text-muted-foreground mb-1.5">
-          These instructions will be prepended to the default PR generation prompt
-        </p>
-        <Textarea
-          className="text-sm min-h-[80px]"
-          placeholder="e.g., Include a testing checklist, link to related issues, use conventional commit format for title"
-          value={prTemplate}
-          onChange={(e) => setPRTemplate(e.target.value)}
-        />
-      </div>
-
-      {hasChanges && (
-        <div className="mt-4 flex justify-end">
-          <Button size="sm" disabled={saving} onClick={handleSave}>
-            {saving ? 'Saving...' : 'Save'}
-          </Button>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function AdvancedSettings() {
-  const [rootDirectory, setRootDirectory] = useState('');
-  const [savedDirectory, setSavedDirectory] = useState('');
-  const [saving, setSaving] = useState(false);
-  const [saveError, setSaveError] = useState<string | null>(null);
-  const minimizeToTray = useSettingsStore((s) => s.minimizeToTray);
-  const setMinimizeToTraySetting = useSettingsStore((s) => s.setMinimizeToTray);
-  const developerMode = useSettingsStore((s) => s.developerMode);
-  const setDeveloperMode = useSettingsStore((s) => s.setDeveloperMode);
-
-  // Fetch the current configured base path from the backend
-  useEffect(() => {
-    getWorkspacesBasePath().then((path) => {
-      setRootDirectory(path);
-      setSavedDirectory(path);
-    }).catch((err) => {
-      console.error('Failed to fetch workspaces base path:', err);
-    });
-  }, []);
-
-  const handleMinimizeToTrayChange = (enabled: boolean) => {
-    setMinimizeToTraySetting(enabled);
-    setMinimizeToTray(enabled); // Sync with Tauri backend
-  };
-
-  const hasUnsavedChanges = rootDirectory !== savedDirectory;
-
-  const handleSaveDirectory = useCallback(async () => {
-    if (!hasUnsavedChanges) return;
-    setSaving(true);
-    setSaveError(null);
-    try {
-      const newPath = await setWorkspacesBasePath(rootDirectory);
-      setSavedDirectory(newPath);
-      setRootDirectory(newPath);
-      // The file watcher will pick up the new path on app restart.
-      // We don't restart it inline because session registrations would be lost.
-    } catch (err) {
-      setSaveError(err instanceof Error ? err.message : 'Failed to save');
-    } finally {
-      setSaving(false);
-    }
-  }, [rootDirectory, hasUnsavedChanges]);
-
-  return (
-    <div>
-      <h2 className="text-xl font-semibold mb-5">Advanced</h2>
-
-      {/* Minimize to tray */}
-      <SettingsRow
-        title="Minimize to tray"
-        description="Keep the app running in the system tray when you close the window"
-      >
-        <Switch
-          checked={minimizeToTray}
-          onCheckedChange={handleMinimizeToTrayChange}
-        />
-      </SettingsRow>
-
-      {/* ChatML root directory */}
-      <div className="py-4 border-b border-border/50">
-        <h4 className="text-sm font-medium">ChatML root directory</h4>
-        <p className="text-sm text-muted-foreground mt-0.5">
-          Where ChatML stores repositories and workspaces. Changing this requires restarting the app.
-        </p>
-        <div className="flex items-center gap-2 mt-3">
-          <input
-            type="text"
-            value={rootDirectory}
-            onChange={(e) => setRootDirectory(e.target.value)}
-            className="flex-1 px-3 py-2 text-sm bg-background border border-border rounded-md focus:outline-none focus:ring-1 focus:ring-primary"
-          />
-          <Button
-            variant="outline"
-            size="sm"
-            className="gap-2 h-9"
-            onClick={async () => {
-              const selectedPath = await openFolderDialog('Select ChatML Root Directory');
-              if (selectedPath) {
-                setRootDirectory(selectedPath);
-              }
-            }}
-          >
-            <FolderOpen className="w-4 h-4" />
-            Browse
-          </Button>
-          {hasUnsavedChanges && (
-            <Button
-              variant="default"
-              size="sm"
-              className="h-9"
-              disabled={saving}
-              onClick={handleSaveDirectory}
-            >
-              {saving ? 'Saving...' : 'Save'}
-            </Button>
-          )}
-        </div>
-        {saveError && (
-          <p className="text-sm text-destructive mt-2">{saveError}</p>
-        )}
-      </div>
-
-      <SettingsRow
-        title="Developer mode"
-        description="Show additional debugging information"
-      >
-        <Switch
-          checked={developerMode}
-          onCheckedChange={setDeveloperMode}
-        />
-      </SettingsRow>
-      <SettingsRow
-        title="Clear cache"
-        description="Clear cached data and temporary files"
-      >
-        <Button variant="outline" size="sm">
-          Clear
-        </Button>
-      </SettingsRow>
     </div>
   );
 }
