+++ src/components/settings/shared/SettingsRow.tsx
@@ -1,0 +1,22 @@
+// Reusable settings row component for label + control layout
+export function SettingsRow({
+  title,
+  description,
+  children,
+}: {
+  title: string;
+  description?: string;
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="flex items-center justify-between py-3 border-b border-border/50">
+      <div className="flex-1 pr-4">
+        <h4 className="text-sm font-medium">{title}</h4>
+        {description && (
+          <p className="text-xs text-muted-foreground mt-0.5">{description}</p>
+        )}
+      </div>
+      <div className="shrink-0">{children}</div>
+    </div>
+  );
+}
